PARTIAL COMMIT - DO NOT MERGE TO MASTER
support for chowdren's new image compression format (extract only, for now)

in addition to having a different metadata block with a different
offset and different amounts of each file, it also uses a new custom
compression format, which is implemented in the new chowimg.py file.

we also still support the old format.