support for chowdren's new image compression format
in addition to having a different metadata block with a different
offset and different amounts of each file, it also uses a new custom
compression format, which is implemented in the new chowimg.py file.

we also still support the old format.  This is done by representing
different asset files with JSON strings that contain metadata such as
metadata block offsets, asset counts and which type of compression to
use for images

* new -r option to dump raw image binaries without decoding them

    this is useless to most users, but it will be helpful for me because i
    can use it for debugging purposes.

    Specifcally, I'm hoping that I can implement compression in a way that
    will produce the exact same binaries as the algorithm used by
    MP2/Galaxy-Trail; however this may be difficult since I do not have
    access to the program used to produce the original Assets.dat.

    The dream is that someday you'll be able to extract Assets.dat with
    the -x option, don't change anything, then recompress it with the -c
    option and get the exact same file as the original input.

* chowimg.py is usable as an independent image extractor/convertor
    * can freely convert between .raw (raw 32-bit RGBA),
      .bin (chowdren's new format), and .png
    * the point of this is so that i can use it for debugging on the compressor.