new -r option to dump raw image binaries without decoding them
this is useless to most users, but it will be helpful for me because i
can use it for debugging purposes.

Specifcally, I'm hoping that I can implement compression in a way that
will produce the exact same binaries as the algorithm used by
MP2/Galaxy-Trail; however this may be difficult since I do not have
access to the program used to produce the original Assets.dat.

The dream is that someday you'll be able to extract Assets.dat with
the -x option, don't change anything, then recompress it with the -c
option and get the exact same file as the original input.