don't add residual data to the last subhunk when compressing
the problem was that the replay data goes *after* the literal data
so residual data needs to be in a new subhunk.

note that if there is no replay data in the last subhunk then we
actually can append residual data to its literal data but currently
it's not implemented that way for the sake of simplicity.